requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# AI and ML dependencies
openai==1.52.0
sentence-transformers==3.2.1
torch==2.8.0
transformers==4.46.3
openai-whisper==20240930

# STT dependencies (pydub removed - Whisper handles formats natively)

# Database
pymysql==1.1.0
sqlalchemy==2.0.23
alembic==1.12.1

# Vector database
chromadb==0.4.18
faiss-cpu==1.7.4

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
numpy==1.24.3
pandas==2.1.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0